Create GlassHouse.java
Once King of Dholokpur has constructed a  glass house for the prince. This glasshouse consists of 18 doors numbering 1, 2, 3 …18. 
On one fine day when prince was playing hide & seek with his friends he went inside this glass house and was struggling to come out of the glass house since all the doors look similar. None other than the king know the solution to this complex problem. 

Based on the location of the prince, the king announces the solution “The sum of the digits of a two digit number, would lead you out of the glass house”


Solution : Write a generic program which accepts a two digit number and find the exit door number of the glasshouse.