simple calculator in java  by switch case
Write a java program to perform simple mathematical operations on two numbers .The user will enter 3 inputs , the first input will be a number ,second input will be an operator and third input will be a number again. It should display the result of the operation as output.
Use switch statement.


solution : Write a program to accept two numbers and an operator to perform mathematical operation on the operands and display the result as output.