Exercise -15 ( Compare two Arrays)
Write a program to find whether 2 arrays are the same.
Input consists of 2n+1 integers. 
The first integer corresponds to ‘n’ , the size of the array.
The next ‘n’ integers correspond to the elements in the first array. 
The next ‘n’ integers correspond to the elements in the second array.
Assume that the maximum value of n is 15.

	
Problem :Write a program to accept values into two separate arrays , then compare the values and print if they are same or different.