Exercise -18 ( Arithmetic Exception)
Write a program that accepts 2 integers a and b as input and finds the quotient of a/b. 
This program may generate an Arithmetic Exception. Use exception handling mechanisms to handle this exception. In the catch block, print the message as shown in the sample output.
 
Also illustrate the use of finally block. Print the message “Inside finally block”.


Solution : Use try, catch and finally to handle ArithmeticException which is raised as a result of division by zero .